% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_model.R
\name{base_model}
\alias{base_model}
\title{Base Model}
\usage{
base_model(Species, R0 = 1e+05, A = 5, MPA = 3, Time1 = 50,
  Time2 = 20, Recruitment_mode = "pool", Error = 0.05,
  Stochasticity = TRUE, Surveys = TRUE, Fishery_management = TRUE,
  Fishing = TRUE, Transects = 24, Adult_movement = TRUE,
  Plotting = FALSE, Final_DR, Years_sampled = 1,
  Areas_sampled = "all", Ind_sampled = "all", Allocation = "IFD",
  BM = FALSE, Control_rules = c(1:6), Output.FM = FALSE,
  Output.N = FALSE, Output.Abundance.All = FALSE,
  Output.Abundance.Mature = FALSE, Output.Biomass = FALSE,
  Output.SSB = FALSE, Output.Catch = FALSE, Output.Yield = FALSE,
  Output.Effort = FALSE, Output.Density.Ratio = FALSE)
}
\arguments{
\item{Species}{character value, the species to analyze. Species names are in
the format species abbreviation.state or stock.year assessed. Example:
BR.CA.2003 is the California black rockfish stock assessed in 2003.}

\item{R0}{numeric value, the unfished recruitment, set arbitrarily. Default
value is 1e+5.}

\item{A}{numeric value, the number of total areas in the model. Default value
is 5.}

\item{MPA}{numeric value, the area number to be designated as a marine
reserve at Time1. Default value is 3.}

\item{Time1}{numeric value, the number of years to run the model before a
marine reserve is implemented. Default value is 50.}

\item{Time2}{numeric value, the number of years to run the model after a
marine reserve is implemented. Default value is 20.}

\item{Recruitment_mode}{character value, values can be:
'closed' - the recruits in each area originate from adults in that area.
'pool' - the recruits in each area come from a pool of larvae produced by
   adults in all areas.
Default value is 'pool'.}

\item{Error}{numeric value, the error between estimated and correct natural
mortality.}

\item{Stochasticity}{logical vector, does recruitment contain a stochastic
component? Default value is TRUE.}

\item{Surveys}{logical value, are surveys being conducted? Default value is
TRUE.}

\item{Fishery_management}{logical value, is the fishery being managed using
the control rules? Default value is TRUE.}

\item{Fishing}{logical value, is fishing occurring? Default value is TRUE.}

\item{Transects}{numerical value, the number of sampling transects conducted
in each area to estimate density ratio. Default value is 24.}

\item{Adult_movement}{logical value, do adults move between areas? Default
value is TRUE.}

\item{Plotting}{logical value, should individual runs be plotted? Default
value is FALSE.}

\item{Final_DR}{numeric value, the final target density ratio.}

\item{Years_sampled}{numeric value, the number of years of sampling upon
which to base the estimate of density ratio. Default value is 1.}

\item{Areas_sampled}{character value, the areas to be sampled to calculate
density ratio. Values can be:
'all' - sample all areas.
'far' - sample only the first and last areas (assuming the reserve is
   somewhere in the middle.
Default value is 'all'.}

\item{Ind_sampled}{character value, the individuals to be sampled to
calculate density ratio. Values can be:
'all' - sample all individuals.
'mature' - sample only mature individuals.
Default value is 'all'.}

\item{Allocation}{character value, how effort is to be allocated. Values can
be:
'IFD' - the ideal free distribution. Effort is allocated proportional to
   the yield caught in each area in the previous timestep.
'equal' - Effort is allocated equally between all areas.
Default value is 'IFD'.}

\item{BM}{logical value, are the control rules from Babcock and MacCall 2011?
Default value is FALSE.}

\item{Control_rules}{numeric vector, the control rules to be compared.
Default value is c(1:6).}

\item{Output.FM}{logical value, should the output include FM? Default value
is FALSE.}

\item{Output.N}{logical value, should the output include N? Default value is
FALSE.}

\item{Output.Abundance.All}{logical value, should the output include
Abundance_all? Default value is FALSE.}

\item{Output.Abundance.Mature}{logical value, should the output include
Abundance_mature? Default value is FALSE.}

\item{Output.Biomass}{logical value, should the output include biomass?
Default value is FALSE.}

\item{Output.SSB}{logical value, should the output include spawning stock
biomass? Default value is FALSE.}

\item{Output.Catch}{logical value, should the output include catch? Default
value is FALSE.}

\item{Output.Yield}{logical value, should the output include yield? Default
value is FALSE.}

\item{Output.Effort}{logical value, should the output include effort? Default
value is FALSE.}

\item{Output.Density.Ratio}{logical value, should the output include density
ratios? Default value is FALSE.}
}
\value{
a list, containing numerical arrays of the relative biomass, yield,
   and spawning stock biomass, as well as the true density ratios over time
   and the transient target density ratios over time
}
\description{
\code{base_model} runs one simulation of the density ratio base model, based
   on Babcock & MacCall (2011).
}
\examples{
base_model(Species = 'BR_CA_2003', R0 = 1e+5, A = 5, MPAs = c(3), Time1 = 50,
   Time2 = 20, Recruitment_mode = 'pool', Error = 0.05, Stochasticity = TRUE,
   Surveys = TRUE, Fishery_management = TRUE, Fishing = TRUE, Transects = 24,
   Adult_movement = TRUE, Plotting = TRUE, Final_DR = 0.6, Years_sampled = 1,
   Areas_sampled = 'all', Ind_sampled = 'all', Allocation = 'IFD',
   Control_rules = c(1:6))
}
