% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sigma.R
\name{compute_sigma}
\alias{compute_sigma}
\title{Compute median interpoint distance sigma}
\usage{
compute_sigma(X, nsamples = nrow(X))
}
\arguments{
\item{X}{A data matrix containing the observations to compute the median interpoint distance on}

\item{nsamples}{The maximum number of samples that is considered}
}
\value{
A scalar \eqn{\sigma} (the median interpoint distance)
}
\description{
Compute median interpoint distance sigma
}
\examples{
x <- rnorm(100) |> matrix(25, 4)
compute_sigma(x)
}
