% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_rule.R
\name{control_rule}
\alias{control_rule}
\title{Control rule}
\usage{
control_rule(t, cr, fdr, A = 5, E, Count, Time1 = 50, TimeT = 70,
  Transects = 24, Nat_mortality, Final_DRs, Inside = 3,
  Outside = c(1, 2, 4, 5), Years_sampled = 1, Areas_sampled = "all",
  Ind_sampled = "all", Floor_DR = 0.2, BM = FALSE,
  Sampling_Error = TRUE, Density_ratio)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{fdr}{temporary numeric value, the current final target density ratio.}

\item{A}{numeric value, the number of total areas in the model. Default
value is 5.}

\item{E}{numeric array, the relative fishing effort displayed in each area,
at each time step, under each control rule, and for each natural mortality
estimate.}

\item{Count}{numeric array, the number of individuals estimated to be in each
area, at each timestep, under each control rule, for each estimate of
natural mortality, for both all individuals and just mature individuals.}

\item{Time1}{numeric value, the number of years to run the model before a
marine reserve is implemented. Default value is 50.}

\item{TimeT}{numeric value, the number of years to run the model total.
Default value is 70.}

\item{Transects}{numerical value, the number of sampling transects conducted
in each area to estimate density ratio. Default value is 24.}

\item{Nat_mortality}{numeric vector, the estimates of natural mortality.}

\item{Final_DRs}{numeric vector, the final target density ratios.}

\item{Inside}{numeric vector, the area(s) inside the marine reserve. Default
is c(3).}

\item{Outside}{numeric vector, the area(s) outside the marine reserve.
Default is c(1, 2, 4, 5).}

\item{Years_sampled}{numeric value, the number of years of sampling upon
which to base the estimate of density ratio. Default value is 1.}

\item{Areas_sampled}{character value, the areas to be sampled to calculate
density ratio. Values can be:
'all' - sample all areas.
'far' - sample only areas farthest from the marine reserve.
Default value is 'all'.}

\item{Ind_sampled}{character value, the individuals to be sampled to
calculate density ratio. Values can be:
'all' - sample all individuals.
'mature' - sample only mature individuals.
Default value is 'all'.}

\item{Floor_DR}{numeric value, the DR value under which effort will be
reduced to 10\% of its starting value. Default value is 0.2.}

\item{BM}{logical value, are the control rules from Babcock and MacCall 2011?
Default value is FALSE.}

\item{Sampling_Error}{logical value, is there any error in sampling? Default
value is TRUE.}

\item{Density_ratio}{numeric array, the true density ratio at each time step
under each control rule and for each final density ratio.}
}
\value{
a numeric vector of fishing effort for the next timestep, under the
   specific control rule, with a specific estimate of natural mortality.
}
\description{
\code{control_rule} determines how to calculate the density ratio and manage
   the fishery (i.e. set fishing effort for the next timestep).
}
\examples{
A = 5; TimeT = 70; CR = 6; NM = 2; FDR = 4; Transects = 24
E <- array(rep(1, A*TimeT*CR*NM*FDR), c(A, TimeT, CR, NM, FDR))
Count <- array(rep(5, A*TimeT*Transects*2*CR*NM*FDR),
   c(A, TimeT, Transects, 2, CR, NM, FDR))
Density_Ratio <- array(rep(0.5, TimeT*CR*FDR), c(TimeT, CR, FDR))
control_rule(t = 51, cr = 1, fdr = 1, A = 5, E, Count, Time1 = 50,
   TimeT = 70, Transects = 24, Nat_mortality = c(0.14, 0.09, 0.19),
   Final_DRs = c(0.2, 0.4, 0.6, 0.8), Inside = 3, Outside = c(1, 2, 4, 5),
   Years_sampled = 1, Areas_sampled = 'all', Ind_sampled = 'all',
   Floor_DR = 0.2, BM = FALSE, Sampling_Error = TRUE, Density_Ratio)
}
