% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_univariate_plot}
\alias{create_univariate_plot}
\title{Indivual univariate plot}
\usage{
create_univariate_plot(data, ext, var, y_lab, sample.facet = TRUE)
}
\arguments{
\item{data}{Data frame with the individual values and density ratio estimates}

\item{ext}{Data frame with the density ratio estimates and sample indicator}

\item{var}{Name of the variable to be plotted on the x-axis}

\item{y_lab}{Name of the y-axis label, typically ("Density Ratio" or "Log Density Ratio")}

\item{sample.facet}{Logical indicating whether to facet the plot by sample. Default is TRUE.}
}
\value{
A scatterplot of variable values and density ratio estimates.
}
\description{
Scatterplot of individual values and density ratio estimates. Used internally in \code{\link[=create_univariate_plot]{create_univariate_plot()}}
}
