% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_ratio.R
\name{density_ratio}
\alias{density_ratio}
\title{Density ratio}
\usage{
density_ratio(t, cr, nm, A, Count, Years_sampled = 1,
  Areas_sampled = "all", Ind_sampled = "all", Transects = 24, Inside,
  Outside)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{nm}{temporary numeric value, the current natural mortality estimate.}

\item{A}{numeric value, the number of total areas in the model. Default
value is 5.}

\item{Count}{numeric array, the number of individuals estimated to be in each
area, at each timestep, under each control rule, for each estimate of
natural mortality, for both all individuals and just mature individuals.}

\item{Years_sampled}{numeric value, the number of years of sampling upon
which to base the estimate of density ratio. Default value is 1.}

\item{Areas_sampled}{character value, the areas to be sampled to calculate
density ratio. Values can be:
'all' - sample all areas.
'far' - sample only the first and last areas (assuming the reserve is
   somewhere in the middle.
Default value is 'all'.}

\item{Ind_sampled}{character value, the individuals to be sampled to
calculate density ratio. Values can be:
'all' - sample all individuals.
'mature' - sample only mature individuals.
Default value is 'all'.}

\item{Transects}{numeric value, the number of sampling transects conducted
in each area to estimate density ratio. Default value is 24.}

\item{Inside}{numeric vector, the area(s) inside the marine reserve. Default
is c(3).}

\item{Outside}{numeric vector, the area(s) outside the marine reserve.
Default is c(1, 2, 4, 5).}
}
\value{
numeric value, the calculated density ratio for a specific timestep,
   under a specific control rule, with a specific estimate of natural
   mortality.
}
\description{
\code{density_ratio} estimates the density ratio for a stock in line with a
   specific control rule.
}
\examples{
Count <- array(rep(5, 5*70*24*2*6*3), c(5, 70, 24, 2, 6, 3))
density_ratio(t = 2, cr = 1, nm = 1, A = 5, Count, Years_sampled = 1,
   Areas_sampled = 'all', Ind_sampled = 'all', Transects = 24,
   Inside = c(3), Outside = c(1, 2, 4, 5))
}
