% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance}
\alias{distance}
\title{Create a Gram matrix with squared Euclidean distances between
observations in the input matrix \code{X} and the input matrix \code{Y}}
\arguments{
\item{X}{A numeric input matrix}

\item{Y}{A numeric input matrix with the same variables as \code{X}}

\item{intercept}{Logical indicating whether an intercept should be added to
the estimation procedure. In this case, the first column is an all-zero
column (which will be transformed into an all-ones column in the kernel).}
}
\description{
Create a Gram matrix with squared Euclidean distances between
observations in the input matrix \code{X} and the input matrix \code{Y}
}
