% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distX}
\alias{distX}
\title{Create a Gram matrix with squared Euclidean distances between all
observations in the input matrix}
\usage{
distX(x, nr, nc)
}
\arguments{
\item{x}{A numeric input matrix}

\item{nr}{Number of rows of the input matrix}

\item{nc}{Number of columns of the input matrix}
}
\description{
Create a Gram matrix with squared Euclidean distances between all
observations in the input matrix
}
