% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distXY}
\alias{distXY}
\title{Create a Gram matrix with squared Euclidean distances between
observations in the input matrix x and the input matrix y}
\usage{
distXY(x, y, nrx, nry, nc)
}
\arguments{
\item{x}{A numeric input matrix}

\item{y}{A numeric input matrix with the same variables as x}

\item{nrx}{Number of rows of the input matrix x}

\item{nry}{Number of rows of the input matrix y}

\item{nc}{Number of columns of the input matrices x and y}
}
\description{
Create a Gram matrix with squared Euclidean distances between
observations in the input matrix x and the input matrix y
}
