% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort_allocation.R
\name{effort_allocation}
\alias{effort_allocation}
\title{Effort allocation}
\usage{
effort_allocation(
  t,
  cr,
  fdr,
  Allocation = "IFD",
  E,
  Yield,
  Time1 = 50,
  Inside = 3,
  Outside = c(1, 2, 4, 5)
)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{fdr}{temporary numeric value, the current final target density ratio.}

\item{Allocation}{character value, how effort is to be allocated. Values can
be:
'IFD' - the ideal free distribution. Effort is allocated proportional to
   the yield caught in each area in the previous timestep.
'equal' - Effort is allocated equally between all areas.
Default value is 'IFD'.}

\item{E}{numeric array, the relative fishing effort displayed in each area,
at each time step, under each control rule, and for each natural mortality
estimate.}

\item{Yield}{numeric array, the yield caught in each area, at each timestep,
under each control rule, and for each estimate of natural mortality, in kg.}

\item{Time1}{numeric value, the number of years to run the model before a
marine reserve is implemented. Default value is 50.}

\item{Inside}{numeric vector, the area(s) inside the marine reserve. Default
value is c(3).}

\item{Outside}{numeric vector, the area(s) outside the marine reserve.
Default value is c(1, 2, 4, 5).}
}
\value{
a numeric array updated with the fishing effort allocation for the
   next timestep.
}
\description{
\code{effort_allocation} redistributes fishing effort for the next timestep.
}
\examples{
A = 5; TimeT = 70; CR = 6; FDR = 4
E <- array(rep(0.2, A*TimeT*CR*FDR), c(A, TimeT, CR, FDR))
Yield <- array(rep(2458, A*TimeT*CR*FDR), c(A, TimeT, CR, FDR))
effort_allocation(t = 51, cr = 1, fdr = 1, Allocation = 'IFD', E,
   Yield, Time1 = 50, Inside = 3, Outside = c(1, 2, 4, 5))
}
