% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_params.kliep}
\alias{extract_params.kliep}
\title{Obtain parameters from a \code{kliep} object}
\usage{
\method{extract_params}{kliep}(object, sigma, ...)
}
\description{
Obtain parameters from a \code{kliep} object
}
\keyword{internal}
