% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_params.kmm}
\alias{extract_params.kmm}
\title{Obtain parameters from a \code{kmm} object}
\usage{
\method{extract_params}{kmm}(object, sigma, ...)
}
\description{
Obtain parameters from a \code{kmm} object
}
\keyword{internal}
