% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_params.lhss}
\alias{extract_params.lhss}
\title{Obtain parameters from a \code{lhss} object}
\usage{
\method{extract_params}{lhss}(object, lambda, lambdasigma, ...)
}
\description{
Obtain parameters from a \code{lhss} object
}
\keyword{internal}
