% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_params.spectral}
\alias{extract_params.spectral}
\title{Obtain parameters from a \code{spectral} object}
\usage{
\method{extract_params}{spectral}(object, sigma, m, ...)
}
\description{
Obtain parameters from a \code{spectral} object
}
\keyword{internal}
