% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{extract_params.ulsif}
\alias{extract_params.ulsif}
\title{Obtain parameters from a \code{ulsif} object}
\usage{
\method{extract_params}{ulsif}(object, sigma, lambda, ...)
}
\description{
Obtain parameters from a \code{ulsif} object
}
\keyword{internal}
