% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_FM.R
\name{historical_FM}
\alias{historical_FM}
\title{Historical fishing mortality to cause estimated depletion}
\usage{
historical_FM(Species, eq_time = 150, R0 = 1e+05, A = 1,
  Stochasticity = FALSE, Recruitment_mode = "pool")
}
\arguments{
\item{Species}{character value, the species to analyze. Species names are in
the format species abbreviation.state or stock.year assessed. Example:
BR.OR.2003 is the Oregon black rockfish stock assessed in 2003.}

\item{eq_time}{numeric value, the number of years to run the function to
determine the historic fishing mortality. Default value is 150.}

\item{R0}{numeric value, set arbitrarily, the unfished recruitment. Default
value is 1e+5.}

\item{A}{numeric value, the number of total areas in the model. Default
value is 5.}

\item{Stochasticity}{logical vector, does recruitment contain a stochastic
component? Default value is FALSE.}

\item{Recruitment_mode}{character value, values can be:
'closed' - the recruits in each area originate from adults in that area.
'pool' - the recruits in each area come from a pool of larvae produced by
   adults in all areas.
Default value is 'pool'.}
}
\value{
a numeric value, the historical value of fishing mortality that would
   result in the estimated depletion value, on the interval (0, 1).
}
\description{
\code{historical_FM} gives the historical fishing mortality that is most
   likely to have resulted in the currently estimated depletion value for the
   fished stock.
}
\examples{
historical_FM(Species = 'BR_CA_2003', eq_time = 150, R0 = 1e+5, A = 1,
   Stochasticity = FALSE, Recruitment_mode = 'pool')
}
