% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{householder_QR}
\alias{householder_QR}
\title{Perform QR decomposition of the input matrix \code{X} through householder
transformations, similar to Matlab's QR decomposition.}
\usage{
householder_QR(X)
}
\arguments{
\item{X}{A numeric matrix.}
}
\description{
Perform QR decomposition of the input matrix \code{X} through householder
transformations, similar to Matlab's QR decomposition.
}
