% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{individual_biv_plot}
\alias{individual_biv_plot}
\title{Bivariate plot}
\usage{
individual_biv_plot(data, vars, logscale, show.sample)
}
\arguments{
\item{data}{Data frame with the individual values and density ratio estimates}

\item{show.sample}{Logical indicating whether to give different shapes to observations, depending on the sample they come from (numerator or denominator). Defaults to FALSE.}
}
\value{
Bivariate plot
}
\description{
Bivariate plot
}
