% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{individual_uni_plot}
\alias{individual_uni_plot}
\title{Indivual univariate plot}
\usage{
individual_uni_plot(data, var, y_lab, sample.facet = TRUE)
}
\arguments{
\item{data}{Data frame with the individual values and density ratio estimates}

\item{var}{Name of the variable to be plotted}

\item{y_lab}{Name of the y-axis label ("Density Ratio" or "Log Density Ratio")}

\item{sample.facet}{Logical indicating whether to facet the plot by sample. Default is TRUE.}
}
\description{
Scatterplot of individual values and density ratio estimates. Used internally in \code{\link[=plot.univariate]{plot.univariate()}}
}
