% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_arrays.R
\name{initialize_arrays}
\alias{initialize_arrays}
\title{Initialize arrays}
\usage{
initialize_arrays(A = 5, Time1 = 50, Time2 = 20, R0 = 1e+05,
  Rec_age, Max_age, A1, L1, A2, L2, K, WA, WB, K_mat, Fb, L50, Sigma_R,
  Rho_R = 0, Fleets, Alpha, A50_up, A50_down, F_fin, Beta, Cf, R, X, SP,
  M, Control_rules, Phi, Stochasticity = TRUE, D, Transects = 24, H,
  Surveys = TRUE, Fishing = TRUE, Error, Recruitment_mode = "pool")
}
\arguments{
\item{A}{numeric value, the number of total areas in the model. Default value
is 5.}

\item{Time1}{numeric value, the number of years to run the model before a
marine reserve is implemented. Default value is 50.}

\item{Time2}{numeric value, the number of years to run the model after a
marine reserve is implemented. Default value is 20.}

\item{R0}{numeric value, the unfished recruitment, set arbitrarily. Default
value is 1e+5.}

\item{Rec_age}{numeric value, the age at recruitment, in years.}

\item{Max_age}{numeric value, the maximum age of fish or total lifespan, in
years.}

\item{A1}{numeric value, age 1 of female fish, in years.}

\item{L1}{numeric value, the length of females at age 1, in cm.}

\item{A2}{numeric value, age 2 of female fish, in years.}

\item{L2}{numeric value, the length of females at age 2, in cm.}

\item{K}{numeric value, the von Bertalanffy growth parameter for females.}

\item{WA}{numeric value, the coefficient in the weight at length equation.}

\item{WB}{numeric value, the exponent in the weight at length equation.}

\item{K_mat}{numeric value, the slope of the maturity curve.}

\item{Fb}{numeric value, the historical fishing effort for the fished species
on the interval (0, 1).}

\item{L50}{numeric value, the length at 50\% maturity, in cm.}

\item{Sigma_R}{numeric value, the recruitment standard deviation.}

\item{Rho_R}{numeric value, the recruitment autocorrelation on the interval
(-1, 1). Default value is 0.}

\item{Fleets}{character vector, the names of all fishing fleets that
contribute to selectivity at age on the interval (0, 1).}

\item{Alpha}{numeric vector, the alpha values for each fishing fleet in
Fleets on the interval (0, 1), gives the relative slope of the upcurve.}

\item{A50_up}{numeric vector, the age values at which selectivity is equal to
0.5 for the upcurve each fishing fleet in Fleets on the interval (0, 1).}

\item{A50_down}{numeric vector, the age values at which selectivity is equal
to 0.5 for the downcurve each fishing fleet in Fleets on the interval
(0, 1).}

\item{F_fin}{numeric vector, the final selectivity values for fleets on the
interval (0, 1).}

\item{Beta}{numeric vector, the beta values for each fishing fleet in Fleets
on the interval (0, 1), gives the relative slope of the downcurve if
dome-shaped.}

\item{Cf}{numeric vector, the fraction of the whole fishery represented by
each fleet.}

\item{R}{numeric value, the proportion of positive transects during sampling.}

\item{X}{numeric value, the average value of individuals seen during positive
transects.}

\item{SP}{numeric value, the standard deviation of individuals seen during
positive transects.}

\item{M}{numeric value, the natural mortality on the interval (0, 1).}

\item{Control_rules}{numeric vector, the control rules to be compared.}

\item{Phi}{numeric value, the unfished recruits per spawner.}

\item{Stochasticity}{logical vector, does recruitment contain a stochastic
component? Default value is TRUE.}

\item{D}{numeric value, the current depletion of the stock, on the interval
(0, 1).}

\item{Transects}{numerical value, the number of sampling transects conducted
in each area to estimate density ratio. Default value is 24.}

\item{H}{numeric value, the steepness of the stock-recruitment curve.}

\item{Surveys}{logical value, are surveys being conducted? Default value is
TRUE.}

\item{Fishing}{logical value, is fishing occurring? Default value is TRUE.}

\item{Error}{numeric value, the error between estimated and correct natural
mortality.}

\item{Recruitment_mode}{character value, values can be:
'closed' - the recruits in each area originate from adults in that area.
'pool' - the recruits in each area come from a pool of larvae produced by
   adults in all areas.
Default value is 'pool'.}
}
\value{
initalizes arrays necessary for other functions in the base model,
   including TimeT, L, W, S, Mat, A50_mat, CR, Nat_mortality, NM, N, SSB,
   Abundance_all, Abundance_mature, Biomass, Eps, B0, Count, Sigma_S, NuS,
   Delta, Gamma, FM, E, Catch, Yield, Rel_biomass, Rel_yield, Rel_SSB,
   Density_ratio
}
\description{
\code{initialize_arrays} returns arrays necessary for other functions in the
   base model: TimeT (total time the model runs), L (length at age),
   W (weight at age), S (selectivity at age), Mat (maturity at age), A50_mat
   (the first age at which 50\% or more of individuals are expected to be
   mature), CR (number of control rules to be compared), Nat_mortality
   (values of estimated natural mortality), NM (number of estimates of
   natural mortality), N (numbers at age array), SSB (spawning stock biomass
   array), Abundance_all (total abundance array), Abundance_mature (mature
   abundance array), Eps (recruitment error), B0 (unfished biomass), Count
   (estimated counts based on sampling array), Sigma_S (sampling random
   variable standard deviation), NuS (sampling random variable), Delta
   (constant of proportionality), Gamma, FM (fishing mortality), E (effort),
   Catch, Yield, Rel_Biomass (relative biomass after reserve implementation),
   Rel_yield (relative yield after reserve implementation), Rel_SSB (relative
   spawning stock biomass after reserve implementation), Density_ratio.
}
\examples{
initialize_arrays(A = 5, Time1 = 50, Time2 = 20, R0 = 1e+5, Rec_age = 2,
   Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15, L2 = 47.95, K = 0.2022,
   WA = 1.68e-5, WB = 3, K_mat = -0.4103, Fb = 0.2, L50 = 39.53,
   Sigma_R = 0.5, Rho_R = 0, Fleets = c('sport', 'hook', 'trawl'),
   Alpha = c(0.33, 0.6, 0.64), A50_up = c(2, 5, 10), A50_down = c(6, 16, 35),
   F_fin = c(0.25, 0.06, 1), Beta = c(1.2, 0.6, 0), Cf = c(0.71, 0.28, 0.01),
   R = 0.77, X = 15.42, SP = 16.97, M = 0.14, Control_rules= c(1:6),
   Phi = 1.1, Stochasticity = TRUE, D = 0.488, Transects = 24, H = 0.65,
   Surveys = TRUE, Fishing = TRUE, Error = 0.05, Recruitment_mode = 'pool')
}
