% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{insurance}
\alias{insurance}
\title{insurance}
\format{
A data.frame with 1338 rows and 7 columns:
\describe{
\item{age}{Age of the insured (continuous)}
\item{sex}{Sex of the insured (binary)}
\item{bmi}{Body mass index of the insured (continuous)}
\item{children}{Number of children/dependents covered by the insurance (integer)}
\item{smoker}{Whether the insured is a smoker (binary)}
\item{region}{The region in which the insured lives (categorical)}
\item{charges}{The medical costs billed by the insurance (continuous)}
}
}
\description{
Insurance data that is openly available (e.g., on
\href{https://www.kaggle.com/datasets/mirichoi0218/insurance}{Kaggle}).
}
\keyword{data}
\keyword{datasets}
\keyword{insurance}
