% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gaussian.R
\name{kernel_gaussian}
\alias{kernel_gaussian}
\title{Compute Gaussian Kernel Gram matrix}
\usage{
kernel_gaussian(x, y = NULL, sigma = NULL)
}
\arguments{
\item{x}{Matrix to compute Gaussian kernel gram matrix with}

\item{y}{\code{NULL} or a matrix with the same variables as \code{x}.
If \code{NULL}, the kernel gram matrix of \code{x} with itself is
computed, otherwise, the kernel gram matrix of \code{x} with
\code{y} is computed.}

\item{sigma}{\code{NULL} or a scalar. If \code{NULL}, sigma is the
median Euclidean interpoint distance.}
}
\value{
The Gaussian kernel gram matrix \eqn{\bf{K}}.
}
\description{
Compute Gaussian Kernel Gram matrix
}
\examples{
x <- rnorm(100) |> matrix(25, 4)
kernel_gaussian(x)
y <- rnorm(200) |> matrix(50, 4)
kernel_gaussian(x, y)
sigma <- 5
kernel_gaussian(x, y, sigma)
}
