% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidiq}
\alias{kidiq}
\title{kidiq}
\format{
A data.frame with 434 rows and 5 columns
\describe{
\item{kid_score}{Child's IQ score (continuous)}
\item{mom_hs}{Whether the mother obtained a high school degree (binary)}
\item{mom_iq}{Mother's IQ score (continuous)}
\item{mom_work}{Whether the mother worked in the first three years of the
child's life (1: not in the first three years; 2: in the second or third
year; 3: parttime in the first year; 4: fulltime in the first year)}
\item{mom_age}{Mother's age (continuous)}
}
}
\description{
The kidiq data stems from the National Longitudinal Survey of Youth and is used
in Gelman and Hill (2007). The data set contains 434 observations measured on
five variables, and is obtained from
\href{https://github.com/jknowles/BDAexampleR}{https://github.com/jknowles/BDAexampleR}.
}
\keyword{data}
\keyword{datasets}
\keyword{kidiq}
