% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_age.R
\name{length_age}
\alias{length_age}
\title{Length at age}
\usage{
length_age(Rec_age, Max_age, A1, L1, A2, L2, K, All_ages = F)
}
\arguments{
\item{Rec_age}{numeric value, the age at recruitment, in years.}

\item{Max_age}{numeric value, the maximum age or total lifespan, in years.}

\item{A1}{numeric value, age 1 of female fish in years.}

\item{L1}{numeric value, the length of females at age 1, in cm.}

\item{A2}{numeric value, age 2 of female fish in years.}

\item{L2}{numeric value, the length of females at age 2, in cm.}

\item{K}{numeric value, the von Bertalanffy growth parameter for females.}

\item{All_ages}{logical value, should the returned vector contain lengths for
ages before recruitment.}
}
\value{
A numeric vector of lengths at ages, from age at recruitment to
   maximum age.
}
\description{
\code{length_age} returns a vector of lengths at ages, in centimeters.
}
\examples{
length_age(Rec_age = 2, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
   L2 = 47.95, K = 0.2022, All_ages = FALSE)
}
