% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_at_age.R
\name{length_at_age}
\alias{length_at_age}
\title{Calculates length at age}
\usage{
length_at_age(rec_age, max_age, L1f, L2f, Kf, a1f, a2f, all_ages = F)
}
\arguments{
\item{rec_age}{numeric value, gives age at which fish has entered the fishery}

\item{max_age}{numeric value}

\item{L1f}{numeric value}

\item{L2f}{numeric value}

\item{Kf}{numeric value}

\item{a1f}{numeric value}

\item{a2f}{numeric value}

\item{all_ages}{logical vector, if T, gives length at ages not included in
population dynamics}
}
\value{
A vector of lengths at ages
}
\description{
Calculates length at age
}
\examples{
length_at_age(3, 40, 17, 49, -0.17, 1, 25, FALSE)
}
