% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management.R
\name{management}
\alias{management}
\title{Management}
\usage{
management(t, cr, fdr, E, DR, target_DR, floor_DR = 0.2,
  effort_inc_allowed = 0.1, Time1 = 50)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{fdr}{temporary numeric value, the current final target density ratio.}

\item{E}{numeric array, the relative fishing effort displayed in each area,
at each time step, under each control rule, and for each natural mortality
estimate.}

\item{DR}{numeric value, the calculated density ratio at a particular
timestep and under a particular control rule.}

\item{target_DR}{numeric value, the target density ratio, below which effort
is allowed to increase, and above which effort is decreased for the next
timestep.}

\item{floor_DR}{numeric value, the density ratio below which effort is set
back to 10\% of the original value. Default value is 0.20.}

\item{effort_inc_allowed}{numeric value, the percent increase allowed in
effort if the density ratio is below the target density ratio, on the
interval (0, 1). Default value is 0.10.}

\item{Time1}{numeric value, the number of years to run the model before a
marine reserve is implemented. Default value is 50.}
}
\value{
a numeric vector of fishing effort for the next timestep, under the
   specific control rule, with a specific estimate of natural mortality.
}
\description{
\code{management} calculates the next timestep's fishing effort allowed,
   given the control rule and estimated density ratio.
}
\examples{
A = 5; TimeT = 70; CR = 6; NM = 3; FDR = 4
E <- array(rep(0.2, A*TimeT*CR*NM*FDR), c(A, TimeT, CR, NM, FDR))
management(t = 51, cr = 1, fdr = 1, E, DR = 0.8, target_DR = 0.6,
   floor_DR = 0.2, effort_inc_allowed = 0.10, Time1 = 50)
}
