% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maturity.R
\name{maturity}
\alias{maturity}
\title{Maturity at age}
\usage{
maturity(Rec_age, Max_age, K_mat, L, L50)
}
\arguments{
\item{Rec_age}{numeric value, the age at recruitment, in years.}

\item{Max_age}{numeric value, the maximum age or total lifespan, in years.}

\item{K_mat}{numeric value, the slope of the maturity curve.}

\item{L}{numeric vector, the length at age vector, in cm.}

\item{L50}{numeric value, the length at 50\% maturity, in cm.}
}
\value{
a numeric vector of fraction mature at ages, from age at recruitment
   to maximum age, on the interval (0, 1).
}
\description{
\code{maturity} calculates the proportion of individuals at each
   age class, from age at recruitment to maximum age, that are expected to be
   mature.
}
\examples{
L <- length_age(Rec_age = 2, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
   L2 = 47.95, K = 0.2022, All_ages = FALSE)
maturity(Rec_age = 2, Max_age = 35, K_mat = -0.4103, L, L50 = 39.53)
}
