% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{naive}
\alias{naive}
\title{Naive density ratio estimation}
\usage{
naive(nu, de, ...)
}
\arguments{
\item{nu}{Numeric matrix with numerator samples}

\item{de}{Numeric matrix with denominator samples (must have the same
variables as \code{nu})}

\item{...}{arguments passed to [stats::density]}
}
\description{
Naive density ratio estimation
}
