% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{naive_subspace}
\alias{naive_subspace}
\title{Naive density ratio estimation}
\usage{
naive_subspace(nu, de, m, ...)
}
\arguments{
\item{nu}{Numeric matrix with numerator samples}

\item{de}{Numeric matrix with denominator samples (must have the same
variables as \code{nu})}

\item{m}{The size (in number of features) of the subspace}

\item{...}{arguments passed to [stats::density]}
}
\description{
Naive density ratio estimation
}
