% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_subspace.R
\name{naivesubspace}
\alias{naivesubspace}
\title{Naive subspace density ratio estimation}
\usage{
naivesubspace(df_numerator, df_denominator, m = NULL, n = 2L^11, ...)
}
\arguments{
\item{df_numerator}{\code{data.frame} with exclusively numeric variables with
the numerator samples}

\item{df_denominator}{\code{data.frame} with exclusively numeric variables
with the denominator samples (must have the same variables as
\code{df_denominator})}

\item{m}{The size (in number of features) of the subspace}

\item{n}{the number of equally spaced points at which the density is to be
estimated. When n > 512, it is rounded up to a power of 2 during the
calculations (as fft is used) and the final result is interpolated by
\link[stats:approxfun]{stats::approx}. So it almost always makes sense to specify n as a power of
two.}

\item{...}{further arguments passed to \link[stats:density]{stats::density}}
}
\description{
The naive subspace estimator first creates an m-dimensional representation
of the data using singular value decomposition, and then runs the \link{naive}
density ratio estimation procedure on the data projected on this subspace.
The SVD is computed using the denominator samples.
}
\examples{
set.seed(456)
# create data that differs only on the first variable
N <- 100
P <- 3 # P-1 noise variables
X <- matrix(rnorm(N*P), N)
Y <- cbind(rnorm(N, 1, 2), matrix(rnorm(N*P-N), N))
df_X <- data.frame(X)
df_Y <- data.frame(Y)

# estimate
dr_naive <- naive(df_X, df_Y)
dr_subspace <- naivesubspace(df_X, df_Y, 1)

# plot: true, naive, naive_subspace
df_new <- data.frame(
  X1 = seq(-4, 4, length.out = 100),
  X2 = seq(-4, 4, length.out = 100),
  X3 = seq(-4, 4, length.out = 100)
)
true_dr <- dnorm(df_new[,1]) / dnorm(df_new[,1], 1, 2)
plot(df_new[,1], true_dr, type = "l", ylab = "Density ratio", ylim = c(0, 3.2))
lines(df_new[,1], predict(dr_naive, df_new), col = "lightblue")
lines(df_new[,1], predict(dr_subspace, df_new), col = "darkorange")

}
