% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{numerator_data}
\alias{numerator_data}
\title{numerator_data}
\format{
A data frame with 1000 rows and 5 columns:
\describe{
\item{x1}{Categorical variable with three categories, 'A', 'B' and 'C'}
\item{x2}{Categorical variable with two categories, 'G1' and 'G2'}
\item{x3}{Continuous variable (normally distributed given x1 and x2)}
\item{x4}{Continuous variable (normally distributed given x3)}
\item{x5}{Continuous variable (mixture of two normally distributed variables)}
}
}
\description{
Simulated data set (see data-raw/generate-data-densityratio.R) with five
variables that are used in the examples.
}
\keyword{data}
\keyword{datasets}
