% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{numerator_small}
\alias{numerator_small}
\title{numerator_small}
\format{
A data frame with 50 rows and 3 columns:
\describe{
\item{x1}{Continuous variable (normally distributed given x1 and x2)}
\item{x2}{Continuous variable (normally distributed given x3)}
\item{x3}{Continuous variable (mixture of two normally distributed variables)}
}
}
\description{
Subset of the \link{numerator_data} with three variables and 50 observations
}
\keyword{data}
\keyword{datasets}
