% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\title{Life history parameters}
\usage{
parameters(Species)
}
\arguments{
\item{Species}{character value, the species to analyze. Species names are in
the format species abbreviation_state or stock_year assessed. Example:
BR.CA.2003 is the California black rockfish stock assessed in 2003.}
}
\value{
a list of numeric values and vectors required to run the base model
   and supporting functions.
}
\description{
\code{parameters} loads the life history parameters for a specific fished
   stock.
}
\examples{
parameters(Species = 'BR_CA_2003')
}
