% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\alias{permute.ulsif}
\title{Single permutation}
\usage{
permute(object, ...)

\method{permute}{ulsif}(object, stacked, nnu, nde)
}
\arguments{
\item{object}{\code{ulsif} object}

\item{...}{Additional arguments to pass through to specific permute functions.}

\item{stacked}{\code{matrix} with stacked numerator and denominator samples}

\item{nnu}{Scalar with numerator sample size}

\item{nde}{Scalar with denominator sample size}
}
\value{
permutation statistic for a single permutation of the data

permutation statistic for a single permutation of the data
}
\description{
Single permutation

Single permutation statistic of \code{ulsif} object
}
