% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.univariate}
\alias{plot.univariate}
\title{Scatter plot of density ratios and individual variables}
\usage{
\method{plot}{univariate}(
  object,
  vars,
  samples = "both",
  logscale = TRUE,
  output = "individual",
  sample.facet = FALSE,
  nrow = NULL
)
}
\arguments{
\item{object}{Density ratio object created with e.g., \code{\link[=kliep]{kliep()}}, \code{\link[=ulsif]{ulsif()}},
or \code{\link[=naive]{naive()}}}

\item{vars}{Character vector of variable names to be plotted.}

\item{logscale}{}

\item{output}{Character indicating whether output should be a list of individual plots ("individual"), or one facetted plot with all variables ("assembled"). Defaults to "individual".}

\item{sample.facet}{Logical indicating whether to facet the plot by sample, i.e, showing plots separate for each sample, and side to side. Defaults to FALSE.}

\item{nrow}{Integer indicating the number of rows in the assembled plot. If NULL, the number of rows is automatically calculated.}
}
\value{
Scatter plot of density ratios and individual variables.
}
\description{
A scatter plot showing the relationship between estimated density ratios and individual variables.
}
