% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_bivariate}
\alias{plot_bivariate}
\title{Title}
\usage{
plot_bivariate(
  object,
  vars,
  samples = "both",
  output = "assembled",
  logscale = TRUE,
  show.sample = FALSE
)
}
\arguments{
\item{show.samples}{}
}
\description{
Title
}
