% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{predict.density}
\alias{predict.density}
\title{Predict function for density ratio estimation}
\usage{
\method{predict}{density}(object, newdata, lambda = 1e-09)
}
\description{
Predict function for density ratio estimation
}
\keyword{internal}
