% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.kliep}
\alias{predict.kliep}
\title{Obtain predicted density ratio values from a \code{kliep} object}
\usage{
\method{predict}{kliep}(object, newdata = NULL, sigma = c("sigmaopt", "all"), ...)
}
\arguments{
\item{object}{A \code{kliep} object}

\item{newdata}{Optional \code{matrix} new data set to compute the density}

\item{sigma}{A scalar with the Gaussian kernel width}

\item{...}{Additional arguments to be passed to the function}
}
\value{
An array with predicted density ratio values from possibly new data,
but otherwise the numerator samples.
}
\description{
Obtain predicted density ratio values from a \code{kliep} object
}
\examples{
x <- rnorm(100) |> matrix(100)
y <- rnorm(200, 1, 2) |> matrix(200)
fit1 <- kliep(x, y)
predict(fit1)
predict(fit1, newdata = rbind(x, y))
predict(fit1, newdata = rbind(x, y), sigma = 2)
}
\seealso{
\code{\link{predict}}, \code{\link{kliep}}
}
\keyword{kliep}
\keyword{predict}
