% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.ulsif}
\alias{predict.ulsif}
\alias{predict.kliep}
\title{Obtain predicted density ratio values from a \code{ulsif} object}
\usage{
\method{predict}{ulsif}(
  object,
  newdata = NULL,
  sigma = c("sigmaopt", "all"),
  lambda = c("lambdaopt", "all"),
  ...
)

\method{predict}{kliep}(object, newdata = NULL, sigma = c("sigmaopt", "all"), ...)
}
\arguments{
\item{object}{Object of class \code{ulsif} or \code{kliep}}
}
\value{
An array with predicted density ratio values from possibly new data,
but otherwise the numerator samples.

An array with predicted density ratio values from possibly new data,
but otherwise the numerator samples.
}
\description{
Obtain predicted density ratio values from a \code{ulsif} object

Obtain predicted density ratio values from a \code{kliep} object
}
