% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lhss}
\alias{print.lhss}
\title{Print a \code{lhss} object}
\usage{
\method{print}{lhss}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{lhss}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{lhss} object.
}
\description{
Print a \code{lhss} object
}
\seealso{
\code{\link{print}}, \code{\link{lhss}}
}
