% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.naivesubspacedensityratio}
\alias{print.naivesubspacedensityratio}
\title{Print a \code{naivedensityratio} object}
\usage{
\method{print}{naivesubspacedensityratio}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{naivesubspacedensityratio}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{naivesubspacedensityratio} object.
}
\description{
Print a \code{naivedensityratio} object
}
\seealso{
\code{\link{print}}, \code{\link{naivesubspace}}
}
