% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ulsif}
\alias{print.ulsif}
\alias{print.kliep}
\title{Print a \code{ulsif} object}
\usage{
\method{print}{ulsif}(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{kliep}(object, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{Object of class \code{ulsif} or \code{kliep}}
}
\value{
\code{NULL}

\code{NULL}
}
\description{
Print a \code{ulsif} object

Print a \code{ulsif} object
}
