% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.spectral}
\alias{print.spectral}
\title{Print a \code{spectral} object}
\usage{
\method{print}{spectral}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{spectral}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{spectral} object.
}
\description{
Print a \code{spectral} object
}
\seealso{
\code{\link{print}}, \code{\link{spectral}}
}
