% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.kliep}
\alias{print.summary.kliep}
\title{Print a \code{summary.kliep} object}
\usage{
\method{print}{summary.kliep}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{summary.kliep}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{summary.kliep} object.
}
\description{
Print a \code{summary.kliep} object
}
\seealso{
\code{\link{print}}, \code{\link{summary.kliep}}, \code{\link{kliep}}
}
