% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.kmm}
\alias{print.summary.kmm}
\title{Print a \code{summary.kmm} object}
\usage{
\method{print}{summary.kmm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{summary.kmm}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{summary.kmm} object.
}
\description{
Print a \code{summary.kmm} object
}
\examples{

\dontshow{
if (requireNamespace("RcppArmadillo", quietly = TRUE)) {
  RcppArmadillo::armadillo_throttle_cores(1)
}
}

set.seed(123)
# Fit model
dr <- kmm(numerator_small, denominator_small)
# Inspect model object
dr
# Obtain summary of model object
summary(dr)
# Plot model object
plot(dr)
# Plot density ratio for each variable individually
plot_univariate(dr)
# Plot density ratio for each pair of variables
plot_bivariate(dr)
# Predict density ratio and inspect first 6 predictions
head(predict(dr))
# Fit model with custom parameters
kmm(numerator_small, denominator_small,
    nsigma = 4, ncenters = 50, nfold = 4,
    constrained = TRUE)

\dontshow{
if(requireNamespace("RcppArmadillo", quietly = TRUE)) {
RcppArmadillo::armadillo_reset_cores()
}
}
}
\seealso{
\code{\link{print}}, \code{\link{summary.kmm}}, \code{\link{kmm}}
}
