% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.naivesubspacedensityratio}
\alias{print.summary.naivesubspacedensityratio}
\title{Print a \code{summary.naivesubspacedensityratio} object}
\usage{
\method{print}{summary.naivesubspacedensityratio}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{summary.naivesubspacedensityratio}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{summary.naivesubspacedensityratio} object.
}
\description{
Print a \code{summary.naivesubspacedensityratio} object
}
\seealso{
\code{\link{print}}, \code{\link{summary.naivesubspacedensityratio}},
\code{\link{naive}}
}
