% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{Recruitment}
\usage{
recruitment(a, t, cr, nm, SSB, A = 5, R0 = 1e+05, H, B0, Eps, Sigma_R,
  Rec_age, Recruitment_mode)
}
\arguments{
\item{a}{temporary numeric value, the current area}

\item{t}{temporary numeric value, the current time step}

\item{cr}{temporary numeric value, the current control rule}

\item{nm}{temporary numeric value, the current natural mortality estimate}

\item{SSB}{numeric matrix, gives the spawning stock biomass of the whole
stock for each area, at each timestep, under each control rule, and for
each estimate of natural mortality}

\item{A}{numeric value, the number of total areas in the model. The default
value is 5.}

\item{R0}{numeric value, set arbitrarily, the unfished recruitment. The
default value is 1e+5.}

\item{H}{numeric value, the steepness of the stock-recruitment curve}

\item{B0}{numeric value, set arbitrarily, the unfished biomass}

\item{Eps}{numeric matrix, recruitment error terms}

\item{Sigma_R}{numeric value, recruitment standard deviation}

\item{Rec_age}{numeric value, gives age at which fish has entered the fishery
in years}

\item{Recruitment_mode}{character value, values can be 'closed' (if the
recruits in each area originate from adults in that area) or 'pool' (if
the recruits in each area come from a pool of larvae produced by all
reproducing individuals in all areas) - the default value is 'pool'}
}
\value{
a numeric value representing the number of new recruits coming into
   the population in area a, at timestep t, under control rule cr, with an
   estimate of natural mortality of nm
}
\description{
\code{recruitment} calculates the number of new recruits entering the
   population
}
\examples{
SSB <- array(rep(10, 5*70*6*3), c(5, 70, 6, 3))
NuR <- array(rnorm(5*70*6*3, 0, 0.5), c(5, 70, 6, 3))
Eps <- epsilon(A = 5, TimeT = 70, CR = 6, NM = 3, NuR, Rho_R = 0)
recruitment(a = 1, t = 3, cr = 1, nm = 2, SSB, A = 5, R0 = 1e+5, H = 0.7,
   B0 = 1e+5/1.1, Eps, Sigma_R = 0.5, Rec_age = 2, Recruitment_mode = 'pool')
}
