% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sampling}
\alias{sampling}
\title{Sampling}
\usage{
sampling(a, t, cr, nm, Delta, Gamma, Abundance_all, Abundance_mature,
  Transects = 24, X, Count, NuS, A = 5)
}
\arguments{
\item{a}{temporary numeric value, the current area.}

\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{nm}{temporary numeric value, the current natural mortality estimate.}

\item{Delta}{numeric value, the proportion of positive transects divided by
depletion, also known as the constant of proportionality}

\item{Gamma}{numeric value, the average value of a positive transect divided
by depletion}

\item{Abundance_all}{numeric array, the total number of individuals in each
area, at each timestep, under all control rules, with all estimates of
natural mortality.}

\item{Abundance_mature}{numeric array, the number of mature individuals in
each area, at each timestep, under all control rules, with all estimates
of natural mortality.}

\item{Transects}{numerical value, the number of sampling transects conducted
in each area to estimate density ratio. Default value is 24.}

\item{X}{numeric value, the average value of individuals seen during positive
transects.}

\item{Count}{numeric array, the number of individuals estimated to be in each
area, at each timestep, under each control rule, for each estimate of
natural mortality, for both all individuals and just mature individuals.}

\item{NuS}{numeric array, the sampling normal variable pulled from a normal
distribution with mean 0 and sd Sigma_S.}

\item{A}{numeric value, the number of total areas in the model. Default value
is 5.}
}
\value{
a two-dimentional numeric array with updated observed counts for the
   fished species in this particular area, at this particular timestep,
   under this particular control rule, with this particular estimate of
   natural mortality.
}
\description{
\code{sampling} samples from the current population to update the Count
   array.
}
\examples{
Abundance_all <- array(rep(340, 5*70*6*3), c(5, 70, 6, 3))
Abundance_mature <- array(rep(280, 5*70*6*3), c(5, 70, 6, 3))
Count <- array(rep(50, 5*70*24*2*6*3), c(5, 70, 24, 2, 6, 3))
NuS <- array(stats::rnorm(5*70*6*3, 0, 0.89), c(5, 70, 6, 3))
sampling(a = 1, t = 51, cr = 1, nm = 1, Delta = 1.6, Gamma = 31.6,
   Abundance_all, Abundance_mature, Transects = 24, X = 15.42, Count, NuS,
   A = 5)
}
