% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_at_age.R
\name{selectivity_at_age}
\alias{selectivity_at_age}
\title{Selectivity at age}
\usage{
selectivity_at_age(Rec_age, Max_age, L, Fleets, A50_up, A50_down, Alpha,
  F_fin, Beta, Cf)
}
\arguments{
\item{Rec_age}{numeric value, gives age at which fish has entered the fishery
in years}

\item{Max_age}{numeric value, gives maximum age of fish or total lifespan in
years}

\item{L}{numeric vector, gives length at age vector, in cm}

\item{Fleets}{character vector, the names of all fishing fleets that
contribute to selectivity at age on the interval (0, 1)}

\item{A50_up}{numeric vector, the age values at which selectivity is equal to
0.5 for the upcurve each fishing fleet in Fleets on the interval (0, 1)}

\item{A50_down}{numeric vector, the age values at which selectivity is equal
to 0.5 for the downcurve each fishing fleet in Fleets on the interval
(0, 1)}

\item{Alpha}{numeric vector, the alpha values for each fishing fleet in
Fleets on the interval (0, 1), gives the relative slope of the upcurve}

\item{F_fin}{numeric vector, the final selectivity values for fleets on the
interval (0, 1)}

\item{Beta}{numeric vector, the beta values for each fishing fleet in Fleets
on the interval (0, 1), gives the relative slope of the downcurve if
dome-shaped}

\item{Cf}{numeric vector, the fraction of the whole fishery represented by
each fleet}
}
\value{
A vector of selectivities at age, from recruitment to maximum age, on
   the interval (0, 1).
}
\description{
\code{selectivity} returns a vector of selectivities at age, on the
   interval (0, 1).
}
\examples{
L <- length_at_age(Rec_age = 2, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
L2 = 47.95, K = 0.2022, All_ages = FALSE)
selectivity(Rec_age = 2, Max_age = 35, L,
Fleets = c('sport', 'hook', 'trawl'), A50_up = c(2, 5, 10),
A50_down = c(6, 16, 35), Alpha = c(0.33, 0.6, 0.64),
F_fin = c(0.25, 0.06, 1), Beta = c(1.2, 0.6, 0), Cf = c(0.71, 0.28, 0.01))
}
