% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stable_AD.R
\name{stable_AD}
\alias{stable_AD}
\title{Stable Age Distribution}
\usage{
stable_AD(Rec_age, Max_age, W, R0, Mat, H, B0, Sigma_R, Fb, S, M, eq_time,
  A50_mat, Stochasticity, Rho_R, Recruitment_mode, LDP = 0.1)
}
\arguments{
\item{Rec_age}{numeric value, the age at recruitment, in years.}

\item{Max_age}{numeric value, the maximum age of fish or total lifespan, in
years.}

\item{W}{numeric vector, the estimated weight at age from age at recruitment
to maximum age, in kg.}

\item{R0}{numeric value, set arbitrarily, the unfished recruitment.}

\item{Mat}{numeric vector, the estimated fraction of individuals mature at
each age, from age at recruitment to maximum age, on the interval (0, 1).}

\item{H}{numeric value, the steepness of the stock-recruitment curve.}

\item{B0}{numeric value, the unfished biomass, in kg.}

\item{Sigma_R}{numeric value, the recruitment standard deviation.}

\item{Fb}{numeric value, the historical fishing effort for the fished species.}

\item{S}{numeric vector, the selectivities at age from age at recruitment to
maximum age, on the interval (0, 1).}

\item{M}{numeric value, the natural mortality on the interval (0, 1).}

\item{eq_time}{numeric value, the number of years to run the function to
determine the stable age distribution.}

\item{A50_mat}{numeric value, the first age at which 50\% or more individuals
are estimated to be mature, on the interval (Rec_age, Max_age).}

\item{Stochasticity}{logical vector, does recruitment contain a stochastic
component? Default value is TRUE.}

\item{Rho_R}{numeric value, the recruitment autocorrelation on the interval
(-1, 1). Default value is 0.}

\item{Recruitment_mode}{character value, values can be:
'closed' - the recruits in each area originate from adults in that area.
'pool' - the recruits in each area come from a pool of larvae produced by
   adults in all areas.
Default value is 'pool'.}

\item{LDP}{numeric value, the larval drift proportion, the proportion of
larvae that drift from one area to an adjacent area before settling.
Default value is 0.1.}
}
\value{
a numeric vector of numbers at age after enough years with no fishing
   that proportions remain stable over time.
}
\description{
\code{stableAD} produces the stable age distribution, or the proportional
   numbers at age from recruitment age to maximum age at equilibrium, after
   enough years of no fishing for proportional numbers at age to stabilize
}
\examples{
L <- length_age(Rec_age = 2, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
   L2 = 47.95, K = 0.2022, All_ages = FALSE)
W <- weight(L, WA = 1.68e-5, WB = 3)
Mat <- maturity(Rec_age = 2, Max_age = 35, K_mat = -0.4103, L, L50 = 39.53)
S <- selectivity(Rec_age = 2, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
   L2 = 47.95, K = 0.2022, Fleets = c('sport', 'hook', 'trawl'),
   A50_up = c(2, 5, 10), A50_down = c(6, 16, 35), Alpha = c(0.33, 0.6, 0.64),
   F_fin = c(0.25, 0.06, 1), Beta = c(1.2, 0.6, 0), Cf = c(0.71, 0.28, 0.01))
stable_AD(Rec_age = 2, Max_age = 35, W, R0 = 1e+5, Mat, H = 0.65, B0 = 1e+5/1.1,
   Sigma_R = 0.5, Fb = 0.2, S, M = 0.14, eq_time = 150, A50_mat = 8,
   Stochasticity = TRUE, Rho_R = 0, Recruitment_mode = 'pool', LDP = 0.1)
}
