% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ulsif}
\alias{summary.ulsif}
\alias{summary.kliep}
\alias{summary.lhss}
\alias{summary.naivedensityratio}
\alias{summary.naivesubspacedensityratio}
\title{Extract summary from \code{ulsif} object, including two-sample significance
test for homogeneity of the numerator and denominator samples}
\usage{
\method{summary}{ulsif}(object, test = FALSE, n.perm = 100, parallel = FALSE, cl = NULL, ...)

\method{summary}{kliep}(object, test = FALSE, n.perm = 100, parallel = FALSE, cl = NULL, ...)

\method{summary}{lhss}(object, test = FALSE, n.perm = 100, parallel = FALSE, cl = NULL, ...)

\method{summary}{naivedensityratio}(object, test = FALSE, n.perm = 100, parallel = FALSE, cl = NULL, ...)

\method{summary}{naivesubspacedensityratio}(object, test = FALSE, n.perm = 100, parallel = FALSE, cl = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{ulsif} or \code{kliep}}

\item{test}{logical indicating whether to statistically test for homogeneity
of the numerator and denominator samples.}

\item{n.perm}{Scalar indicating number of permutation samples}

\item{parallel}{\code{logical} indicating to run the permutation test in parallel}

\item{cl}{\code{NULL} or a cluster object created by \code{makeCluster}. If
\code{NULL} and \code{parallel = TRUE}, it uses the number of available cores
minus 1.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Summary of the fitted density ratio model

Summary of the fitted density ratio model

Summary of the fitted density ratio model

Summary of the fitted density ratio model

Summary of the fitted density ratio model
}
\description{
Extract summary from \code{ulsif} object, including two-sample significance
test for homogeneity of the numerator and denominator samples

Extract summary from \code{kliep} object, including two-sample significance
test for homogeneity of the numerator and denominator samples

Extract summary from \code{lhss} object, including two-sample significance
test for homogeneity of the numerator and denominator samples

Extract summary from \code{naivedensityratio} object, including two-sample
significance test for homogeneity of the numerator and denominator samples

Extract summary from \code{naivesubspacedensityratio} object, including
two-sample significance test for homogeneity of the numerator and
denominator samples
}
