% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient_DR.R
\name{transient_DR}
\alias{transient_DR}
\title{Transient DR}
\usage{
transient_DR(Time1 = 50, TimeT = 70, Final_DRs, M, fdr)
}
\arguments{
\item{Time1}{numeric value, the number of years to run the model before a
marine reserve is implemented. Default value is 50.}

\item{TimeT}{numeric value, the number of years to run the model total.
Default value is 70.}

\item{Final_DRs}{numeric vector, the final target density ratios.}

\item{M}{numeric value, the natural mortality on the interval (0, 1).}

\item{fdr}{temporary numeric value, the current final target density ratio.}
}
\value{
a numeric vector, the target density ratios over all timesteps with
   transient control rules
}
\description{
\code{transient_DR} calculates the target density ratios, over all timesteps
   with management, assuming a transient control rule is being used.
}
\examples{
transient_DR(Time1 = 50, TimeT = 70, Final_DRs = c(0.6, 0.8), M = 0.17,
   fdr = 1)
}
