% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_DR.R
\name{true_DR}
\alias{true_DR}
\title{True density ratio}
\usage{
true_DR(t, cr, fdr, Abundance, Inside = 3, Outside = c(1, 2, 4, 5),
  Density_ratio)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{fdr}{temporary numeric value, the current final target density ratio.}

\item{Abundance}{numeric array, the total number of individuals in each area,
at each timestep, under all control rules, with all estimates of natural
mortality.}

\item{Inside}{numeric vector, the area(s) inside the marine reserve. Default
value is 3.}

\item{Outside}{numeric vector, the area(s) outside the marine reserve.
Default value is c(1, 2, 4, 5).}

\item{Density_ratio}{numeric array, the true density ratios at each timestep,
under each control rule, and for each estimate of natural mortality.}
}
\value{
a numeric array that updates the true density ratios for the most
   current timestep and control rule.
}
\description{
\code{true_DR} returns the true density ratio for the current timestep,
   updating the Density_ratio array.
}
\examples{
A = 5; TimeT = 70; CR = 6; NM = 1; FDR = 4; Time2= 20
Abundance <- array(rep(3400, A*TimeT*CR*FDR*1*NM),
c(A, TimeT, CR, FDR, 1, NM))
Density_ratio <- array(rep(0, TimeT*CR*FDR), c(TimeT, CR, FDR))
true_DR(t = 51, cr = 1, fdr = 1, Abundance, Inside = 3,
   Outside = c(1, 2, 4, 5), Density_ratio)
}
