% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulsif.R
\name{ulsif}
\alias{ulsif}
\title{Unconstrained least-squares importance fitting}
\usage{
ulsif(nu, de, sigma = NULL, lambda = NULL, maxcenters = nrow(nu))
}
\arguments{
\item{nu}{Numeric matrix with numerator samples}

\item{de}{Numeric matrix with denominator samples (must have the same
variables as \code{nu})}

\item{sigma}{\code{NULL} or a scalar value to determine the bandwidth of the
Gaussian kernel gram matrix. If \code{NULL}, sigma is the median Euclidean
interpoint distance.}

\item{lambda}{\code{NULL} or a scalar value to determine the regularization
imposed on the Gaussian kernel gram matrix of the denominator samples. If
\code{NULL}, \code{lambda} is chosen to be \eqn{\sqrt{N}}.}
}
\value{
\code{kmm} returns \code{rhat_de}, the estimated density ratio for
the denominator samples.
}
\description{
Unconstrained least-squares importance fitting
}
\examples{
x <- rnorm(100) |> matrix(100)
y <- rnorm(200, 1, 2) |> matrix(200)
kmm(x, y)
kmm(x, y, sigma = 2, lambda = 2)

}
