% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_at_age.R
\name{weight_at_age}
\alias{weight_at_age}
\title{Weight at age}
\usage{
weight_at_age(L, WA, WB)
}
\arguments{
\item{L}{numeric vector, gives length at age vector, in cm}

\item{WA}{numeric value, coefficient in weight at length equation, where
weight = WA * length ^ WB}

\item{WB}{numeric value, exponent in weight at length equation, where
weight = WA * length ^ WB}
}
\value{
a numeric vector of weights at ages, from age at recruitment to
   maximum age
}
\description{
\code{weight_at_age} returns a vector of weights at age, in kg.
}
\examples{
L <- length_at_age(Rec_age = 2, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
L2 = 47.95, K = 0.2022, All_ages = FALSE)
weight_at_age(L, WA = 1.68e-5, WB = 3)
}
