% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch.R
\name{catch}
\alias{catch}
\title{Catch at age}
\usage{
catch(t, cr, nm, fdr, FM, Nat_mortality, N, A, Fb, E, Catch)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{nm}{temporary numeric value, the current natural mortality estimate.}

\item{fdr}{temporary numeric value, the current final target density ratio.}

\item{FM}{numeric array that corresponds to the fishing mortality at each
age in each area, at each timestep, under all control rules, with all
estimates of natural mortality.}

\item{Nat_mortality}{numeric vector, the estimates of natural mortality.}

\item{N}{numeric array, the number of individuals at each age, in each
area, at each timestep, under each control rule, and for each estimate of
natural mortality.}

\item{A}{numeric value, the number of total areas in the model. Default
value is 5.}

\item{Fb}{numeric value, the historical fishing effort for the fished species.}

\item{E}{numeric array, the relative fishing effort displayed in each area,
at each time step, under each control rule, and for each natural mortality
estimate.}

\item{Catch}{numeric array, the number of individuals caught at each age, in
each area, at each timestep, under each control rule, for each estimate of
natural mortality.}
}
\value{
a numeric array with an updated vector of catch at ages given the
   current area, timestep, control rule, and estimate of natural mortality.
}
\description{
\code{catch} returns the catch at age vector. It is based on the continuous
(Baranov) formulation of of selectivity, such that fishing mortality (FM) and
natural mortality (M) are proportional to fish abundance (N) and act
simultaneously  and uniformly throughout the year, i.e. dN/dt = -(M+F)*N.
}
\examples{
n = 34; A = 5; TimeT = 70; CR = 6; NM = 3; FDR = 4
FM <- array(rep(0.2, n*A*TimeT*CR*NM*FDR), c(n, A, TimeT, CR, NM, FDR))
N <- array(rep(10, n*A*TimeT*CR*NM*FDR), c(n, A, TimeT, CR, NM, FDR))
E <- array(rep(1, A*TimeT*CR*NM*FDR), c(A, TimeT, CR, NM, FDR))
Catch <- array(rep(2, n*A*TimeT*CR*NM*FDR), c(n, A, TimeT, CR, NM, FDR))
catch(t = 1, cr = 1, nm = 1, fdr = 1, FM,
   Nat_mortality = c(0.09, 0.14, 0.19), N, A = 5, Fb = 0.2, E, Catch)
}
