% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon}
\alias{colon}
\title{colon}
\format{
A data.frame with 62 rows and 2001 columns (class variable and 2000
gene expressions).
}
\description{
Colon cancer data set from princeton, containing 2000 gene expressions from
22 colon tumor tissues and 40 non-tumor tissues. The data can also be found
\href{here}{http://genomics-pubs.princeton.edu/oncology/affydata/I2000.html}.
}
\references{
Alon, U., Barkai, N., Notterman, D. A., Gish, K., Ybarra, S., Mack, D., &
Levine, A. J. (1999). Broad patterns of gene expression revealed by clustering
of tumor and normal colon tissues probed by oligonucleotide arrays.
\emph{Proceedings of the National Academy of Sciences}, 96(12), 6745-6750.
}
\keyword{colon}
\keyword{data}
\keyword{datasets}
