% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon}
\alias{colon}
\title{colon}
\format{
A data.frame with 62 rows and 2001 columns (class variable and 2000
gene expressions).
}
\description{
Colon cancer data set from princeton, containing 2000 gene expressions from
22 colon tumor tissues and 40 non-tumor tissues. The data is collected by
Alon et al. (1999) and can be obtained from
\href{http://genomics-pubs.princeton.edu/oncology/affydata/}{here}.
}
\keyword{colon}
\keyword{data}
\keyword{datasets}
