% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_rule.R
\name{control_rule}
\alias{control_rule}
\title{Control rule}
\usage{
control_rule(t, cr, A = 5, E, Count, Time1 = 50, TimeT = 70,
  Transects = 24, Nat_mortality, Final_DR, Inside = c(3),
  Outside = c(1, 2, 4, 5), Areas_sampled = "all",
  Ind_sampled = "all", Years_sampled = 1)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{A}{numeric value, the number of total areas in the model. Default
value is 5.}

\item{E}{numeric array, the relative fishing effort displayed in each area,
at each time step, under each control rule, and for each natural mortality
estimate.}

\item{Count}{numeric array, the number of individuals estimated to be in each
area, at each timestep, under each control rule, for each estimate of
natural mortality, for both all individuals and just mature individuals.}

\item{Time1}{numeric value, the number of years to run the model before a
marine reserve is implemented. Default value is 50.}

\item{TimeT}{numeric value, the number of years to run the model total.
Default value is 70.}

\item{Transects}{numerical value, the number of sampling transects conducted
in each area to estimate density ratio. Default value is 24.}

\item{Nat_mortality}{numeric vector, the estimates of natural mortality.}

\item{Final_DR}{numeric value, the final target density ratio.}

\item{Inside}{numeric vector, the area(s) inside the marine reserve. Default
is c(3).}

\item{Outside}{numeric vector, the area(s) outside the marine reserve.
Default is c(1, 2, 4, 5).}

\item{Areas_sampled}{character value, the areas to be sampled to calculate
density ratio. Values can be:
'all' - sample all areas.
'far' - sample only areas farthest from the marine reserve.
Default value is 'all'.}

\item{Ind_sampled}{character value, the individuals to be sampled to
calculate density ratio. Values can be:
'all' - sample all individuals.
'mature' - sample only mature individuals.
Default value is 'all'.}

\item{Years_sampled}{numeric value, the number of years of sampling upon
which to base the estimate of density ratio. Default value is 1.}
}
\value{
a numeric vector of fishing effort for the next timestep, under the
   specific control rule, with a specific estimate of natural mortality.
}
\description{
\code{control_rule} determines how to calculate the density ratio and manage
   the fishery (i.e. set fishing effort for the next timestep).
}
\examples{
E <- array(rep(1, 5*70*6*3), c(5, 70, 6, 3))
Count <- array(rep(5, 5*70*24*2*6*3), c(5, 70, 24, 2, 6, 3))
control_rule(t = 51, cr = 1, A = 5, E, Count, Time1 = 50, TimeT = 70,
   Transects = 24, Nat_mortality = c(0.09, 0.14, 0.19), Final_DR = 0.6,
   Inside = c(3), Outside = c(1, 2, 4, 5), Areas_sampled = 'all',
   Ind_sampled = 'all', Years_sampled = 1)
}
