% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{denominator_small}
\alias{denominator_small}
\title{denominator_small}
\format{
A data frame with 100 rows and 3 columns:
\describe{
\item{x1}{Continuous variable (normally distributed given x1 and x2)}
\item{x2}{Continuous variable (normally distributed)}
\item{x3}{Continuous variable (normally distributed)}
}
}
\description{
Subset of the \link{denominator_data} with three variables and 50 observations
}
\keyword{data}
\keyword{datasets}
