% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epsilon.R
\name{epsilon}
\alias{epsilon}
\title{Epsilon}
\usage{
epsilon(A = 5, TimeT = 70, CR = 6, NM = 3, FDR, NuR, Rho_R = 0)
}
\arguments{
\item{A}{numeric value, the number of total areas in the model. Default value
is 5.}

\item{TimeT}{numeric value, the number of years to run the model total.
Default value is 70.}

\item{CR}{numeric value, the number of control rules to be compared. Default
value is 6.}

\item{NM}{numeric value, the total number of estimated values of natural
mortality. Default value is 3.}

\item{FDR}{numeric value, the total number of final density ratios.}

\item{NuR}{numeric vector, the recruitment random normal variable, pulled
from a normal distribution of mean 0 and standard deviation equal to
Sigma_R.}

\item{Rho_R}{numeric value, the recruitment autocorrelation on the interval
(-1, 1). Default value is 0.}
}
\value{
a numeric vector of recruitment error terms, of dimensions
   A \* timeT \* CR \* NM.
}
\description{
\code{epsilon} returns a vector of error terms for recruitment, based on
equation 4 in Babcock & MacCall (2011).
}
\examples{
A = 5; TimeT = 70; CR = 6; NM = 3; FDR = 4
NuR <- array(stats::rnorm(A*TimeT*CR*NM*FDR, 0, 0.5), c(A, TimeT, CR, NM, FDR))
epsilon(A, TimeT, CR, NM, FDR, NuR, Rho_R = 0)
}
