% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraction_mature_at_age.R
\name{frac_mat}
\alias{frac_mat}
\title{Fraction mature at age}
\usage{
frac_mat(Rec_age, Max_age, K_mat, L, L50)
}
\arguments{
\item{Rec_age}{numeric value, gives age at which fish has entered the fishery
in years}

\item{Max_age}{numeric value, gives maximum age of fish or total lifespan in
years}

\item{K_mat}{numeric value, slope of maturity curve}

\item{L}{numeric vector, gives length at age vector, in cm}

\item{L50}{numeric value, gives length at 50% maturity, in cm}
}
\value{
a numeric vector of fractions mature at age, or the proportion of
   individuals that are expected to be mature at any age, on the interval
   (0, 1).
}
\description{
Fraction mature at age
}
\examples{
L <- length_at_age(Rec_age = 2, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
L2 = 47.95, K = 0.2022, All_ages = FALSE)
frac_mat(Rec_age = 3, Max_age = 35, K_mat = -0.4103, L, L50 = 39.53)
}
