% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_gaussian.R
\name{kernel_gaussian}
\alias{kernel_gaussian}
\title{Compute Gaussian Kernel Gram matrix}
\usage{
kernel_gaussian(distance, sigma = NULL, symmetric = FALSE)
}
\arguments{
\item{distance}{Numeric distance matrix computed with \code{distance}}

\item{sigma}{\code{NULL} or a scalar. If \code{NULL}, sigma is the
median Euclidean interpoint distance.}

\item{symmetric}{Logical indicating whether the distance matrix is symmetric
which can speed up calculations.}
}
\value{
The Gaussian kernel gram matrix \eqn{\bf{K}}.
}
\description{
Compute Gaussian Kernel Gram matrix
}
\examples{
x <- rnorm(100) |> matrix(25, 4)
densityratio:::distance(x, x, TRUE) |> densityratio:::kernel_gaussian(symmetric = TRUE)
y <- rnorm(200) |> matrix(50, 4)
densityratio:::distance(x, y) |> densityratio:::kernel_gaussian()
sigma <- 5
densityratio:::distance(x, y) |> densityratio:::kernel_gaussian(sigma)
}
