% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kernel_gaussian}
\alias{kernel_gaussian}
\title{Create gaussian kernel gram matrix from distance matrix}
\arguments{
\item{dist}{A numeric distance matrix}

\item{sigma}{A scalar with the length-scale parameter}
}
\description{
Create gaussian kernel gram matrix from distance matrix
}
