% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kliep.R
\name{kliep}
\alias{kliep}
\title{Kullback-Leibler importance estimation procedure}
\usage{
kliep(
  nu,
  de,
  sigma = NULL,
  maxcenters = nrow(nu),
  centers = NULL,
  eps = 0.001,
  maxit = 100,
  printFlag = T
)
}
\arguments{
\item{nu}{Numeric matrix with numerator samples}

\item{de}{Numeric matrix with denominator samples (must have the same
variables as \code{nu})}

\item{sigma}{\code{NULL} or a scalar value to determine the bandwidth of the
Gaussian kernel gram matrix. If \code{NULL}, sigma is the median Euclidean
interpoint distance.}

\item{maxcenters}{Maximum number of Gaussian centers in the kernel gram
matrix. Defaults to all numerator samples.}

\item{centers}{Option to specify the Gaussian samples manually.}

\item{eps}{Learning rate for the optimization procedure (can be a vector)}

\item{maxit}{Maximum number of iterations for the optimization scheme.}

\item{printFlag}{Whether or not to print the progress of the optimization
scheme.}
}
\value{
\code{kliep} returns \code{rhat_de}, the estimated density ratio for
the denominator samples.
}
\description{
Kullback-Leibler importance estimation procedure
}
\examples{
x <- rnorm(100) |> matrix(100)
y <- rnorm(200, 1, 2) |> matrix(200)
kliep(x, y)
kliep(x, y, sigma = 2)

}
