% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmm.R
\name{kmm}
\alias{kmm}
\title{Kernel mean matching approach to density ratio estimation}
\usage{
kmm(nu, de, method = "unconstrained", sigma = NULL, lambda = NULL)
}
\arguments{
\item{nu}{Numeric matrix with numerator samples}

\item{de}{Numeric matrix with denominator samples (must have the same
variables as \code{nu})}

\item{method}{Character string containing the method used for kernel mean
matching. Currently, \code{method = "unconstrained"} and
\code{method = "constrained"} are supported.}

\item{sigma}{\code{NULL} or a scalar value to determine the bandwidth of the
Gaussian kernel gram matrix. If \code{NULL}, \code{sigma} is the median Euclidean
interpoint distance.}

\item{lambda}{\code{NULL} or a scalar value to determine the regularization
imposed on the Gaussian kernel gram matrix of the denominator samples. If
\code{NULL}, \code{lambda} is chosen to be \eqn{\sqrt{N}}.}
}
\value{
\code{kmm} returns \code{rhat_de}, the estimated density ratio for
the denominator samples.
}
\description{
Kernel mean matching approach to density ratio estimation
}
\examples{
x <- rnorm(100) |> matrix(100)
y <- rnorm(200, 1, 2) |> matrix(200)
kmm(x, y)
kmm(x, y, sigma = 2, lambda = 2)

}
