% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_at_age.R
\name{length_at_age}
\alias{length_at_age}
\title{Length at age}
\usage{
length_at_age(Rec_age, Max_age, A1, L1, A2, L2, K, All_ages = F)
}
\arguments{
\item{Rec_age}{numeric, gives age at which fish has entered the fishery in years}

\item{Max_age}{numeric, gives maximum age of fish or total lifespan in years}

\item{A1}{numeric, age 1 of female fish in years}

\item{L1}{numeric, gives the length of females at age 1 in cm}

\item{A2}{numeric, age 2 of female fish in years}

\item{L2}{numeric, gives the length of females at age 2 in cm}

\item{K}{numeric, gives the von Bertalanffy growth parameter for females}

\item{All_ages}{logical, determines if the returned vector should contain lengths for ages before recruitment}
}
\value{
A vector of lengths at ages, from rec_age to max_age
}
\description{
\code{length_at_age} returns a vector of lengths at ages, in centimeters
}
\examples{
length_at_age(Rec_age = 3, Max_age = 35, A1 = 5, L1 = 32.21, A2 = 15,
L2 = 47.95, K = 0.2022, All_ages = FALSE)
}
