% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhss.R
\name{lhss}
\alias{lhss}
\title{Least-squares heterodistributional subspace search}
\usage{
lhss(
  df_numerator,
  df_denominator,
  m = 1,
  nsigma = 10,
  sigma_quantile = NULL,
  sigma = NULL,
  nlambda = 10,
  lambda = NULL,
  ncenters = 200,
  centers = NULL,
  maxit = 200,
  parallel = FALSE,
  nthreads = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{df_numerator}{\code{data.frame} with exclusively numeric variables with
the numerator samples}

\item{df_denominator}{\code{data.frame} with exclusively numeric variables
with the denominator samples (must have the same variables as
\code{df_denominator})}

\item{m}{Scalar indicating the dimensionality of the reduced subspace}

\item{sigma}{\code{NULL} or a scalar value to determine the bandwidth of the
Gaussian kernel gram matrix. If \code{NULL}, sigma is the median Euclidean
interpoint distance.}

\item{lambda}{\code{NULL} or a scalar value to determine the regularization
imposed on the Gaussian kernel gram matrix of the denominator samples. If
\code{NULL}, \code{lambda} is chosen to be \eqn{\sqrt{N}}.}

\item{ncenters}{Maximum number of Gaussian centers in the kernel gram
matrix. Defaults to all numerator samples.}

\item{centers}{Numeric matrix with the same variables as \code{nu} and
\code{de} that are used as Gaussian centers in the kernel Gram matrix. By
default, the matrix \code{nu} is used as the matrix with Gaussian centers.}

\item{maxit}{Maximum number of iterations in the updating scheme.}
}
\value{
\code{lhss} returns \code{rhat}, the estimated density ratio.
}
\description{
Least-squares heterodistributional subspace search
}
\examples{
\dontrun{
set.seed(1)
N <- 1000
X <- cbind(rnorm(N), rnorm(N, 0, 0.5))
Y <- cbind(rnorm(N), sample(rep(c(-1, 1), times = N/2)) + rnorm(N))
out <- lhss(X, Y, m = 1, ncenters = 100)
}
}
