% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhss.R
\name{lhss}
\alias{lhss}
\title{Least-squares heterodistributional subspace search}
\usage{
lhss(
  df_numerator,
  df_denominator,
  m = NULL,
  intercept = TRUE,
  nsigma = 10,
  sigma_quantile = NULL,
  sigma = NULL,
  nlambda = 10,
  lambda = NULL,
  ncenters = 200,
  centers = NULL,
  maxit = 200,
  parallel = FALSE,
  nthreads = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{df_numerator}{\code{data.frame} with exclusively numeric variables with
the numerator samples}

\item{df_denominator}{\code{data.frame} with exclusively numeric variables
with the denominator samples (must have the same variables as
\code{df_denominator})}

\item{m}{Scalar indicating the dimensionality of the reduced subspace}

\item{intercept}{\code{logical} Indicating whether to include an intercept
term in the model. Defaults to \code{TRUE}.}

\item{nsigma}{Integer indicating the number of sigma values (bandwidth
parameter of the Gaussian kernel gram matrix) to use in cross-validation.}

\item{sigma_quantile}{\code{NULL} or numeric vector with probabilities to
calculate the quantiles of the distance matrix to obtain sigma values. If
\code{NULL}, \code{nsigma} values between \code{0.25} and \code{0.75} are
used.}

\item{sigma}{\code{NULL} or a scalar value to determine the bandwidth of the
Gaussian kernel gram matrix. If \code{NULL}, \code{nsigma} values between
\code{0.25} and \code{0.75} are used.}

\item{nlambda}{Integer indicating the number of \code{lambda} values
(regularization parameter), by default, \code{lambda} is set to
\code{10^seq(3, -3, length.out = nlambda)}.}

\item{lambda}{\code{NULL} or numeric vector indicating the lambda values to
use in cross-validation}

\item{ncenters}{Maximum number of Gaussian centers in the kernel gram
matrix. Defaults to all numerator samples.}

\item{centers}{Numeric matrix with the same variables as \code{nu} and
\code{de} that are used as Gaussian centers in the kernel Gram matrix. By
default, the matrix \code{nu} is used as the matrix with Gaussian centers.}

\item{maxit}{Maximum number of iterations in the updating scheme.}

\item{parallel}{logical indicating whether to use parallel processing in the
cross-validation scheme.}

\item{nthreads}{\code{NULL} or integer indicating the number of threads to
use for parallel processing. If parallel processing is enabled, it defaults
to the number of available threads minus one.}

\item{progressbar}{Logical indicating whether or not to display a progressbar.}
}
\value{
\code{lhss}-object, containing all information to calculate the
density ratio using optimal sigma, optimal lambda and optimal weights.

\code{lhss} returns \code{rhat}, the estimated density ratio.
}
\description{
Least-squares heterodistributional subspace search
}
\examples{
set.seed(1)
N <- 250
X <- cbind(rnorm(N), rnorm(N, 0, 0.5))
Y <- cbind(rnorm(N), sample(rep(c(-1, 1), times = N/2)) + rnorm(N))
out <- lhss(X, Y, m = 1, ncenters = 100)
}
