% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movement.R
\name{movement}
\alias{movement}
\title{Movement}
\usage{
movement(t, cr, fdr, N, A, AMP = 0.1)
}
\arguments{
\item{t}{temporary numeric value, the current time step.}

\item{cr}{temporary numeric value, the current control rule.}

\item{fdr}{temporary numeric value, the current final target density ratio.}

\item{N}{numeric array, the number of individuals at each age, in each
area, at each timestep, under each control rule, and for each estimate of
natural mortality.}

\item{A}{numeric value, the number of total areas in the model. Default
value is 5.}

\item{AMP}{numeric value, adult movement proportion, the fraction of
individuals that move from one area to an adjacent area from one timestep
to the next. Default value is 0.1.}
}
\value{
numeric array of updated N (numbers at age, area, timestep, control
   rule, and estimate of natural mortality)
}
\description{
\code{movement} redistributes individuals amongst areas.
}
\examples{
A = 5; TimeT = 70; CR = 6; FDR = 4
N <- array(rep(10, n*A*TimeT*CR*FDR), c(n, A, TimeT, CR, FDR))
movement(t = 1, cr = 1, fdr = 1, N, A = 5, AMP = 0.1)
}
