% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{naive}
\alias{naive}
\title{Naive density ratio estimation}
\usage{
naive(df_numerator, df_denominator, n = 2L^11, ...)
}
\arguments{
\item{df_numerator}{\code{data.frame} with exclusively numeric variables with
the numerator samples}

\item{df_denominator}{\code{data.frame} with exclusively numeric variables
with the denominator samples (must have the same variables as
\code{df_denominator})}

\item{n}{\code{integer} the number of equally spaced points at which the density is
estimated. When n > 512, it is rounded up to a power of 2 during the
calculations (as fft is used) and the final result is interpolated by
\link[stats:approxfun]{stats::approx}. So it makes sense to specify n as a power of two.}

\item{...}{further arguments passed to \link[stats:density]{stats::density}}
}
\value{
\code{naivedensityratio} object
}
\description{
The naive approach creates separate kernel density estimates for
the numerator and the denominator samples, and then evaluates their
ratio for the denominator samples. For multivariate data, this
approach assumes the variables are independent (naive Bayes assumption).
}
\examples{
x <- rnorm(100)
y <- rnorm(200, 1, 2)

naive(x, y)
naive(x, y, bw = 2)

}
\seealso{
\code{\link[stats:density]{stats::density()}}
}
