% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\alias{permute.ulsif}
\alias{permute.kliep}
\alias{permute.lhss}
\alias{permute.spectral}
\alias{permute.naivedensityratio}
\title{Single permutation}
\usage{
permute(object, ...)

\method{permute}{ulsif}(object, stacked, nnu, nde, ...)

\method{permute}{kliep}(object, stacked, nnu, nde, min_pred = sqrt(.Machine$double.eps), ...)

\method{permute}{lhss}(object, stacked, nnu, nde, ...)

\method{permute}{spectral}(object, stacked, nnu, nde, ...)

\method{permute}{naivedensityratio}(object, stacked, nnu, nde, min_pred, max_pred)
}
\arguments{
\item{object}{\code{naivedensityratio} object}

\item{...}{Additional arguments to pass through to specific permute functions.}

\item{stacked}{\code{matrix} with stacked numerator and denominator samples}

\item{nnu}{Scalar with numerator sample size}

\item{nde}{Scalar with denominator sample size}

\item{min_pred}{Minimum value of the predicted density ratio}

\item{max_pred}{Maximum value of the predicted density ratio}
}
\value{
permutation statistic for a single permutation of the data

permutation statistic for a single permutation of the data

permutation statistic for a single permutation of the data

permutation statistic for a single permutation of the data

permutation statistic for a single permutation of the data

permutation statistic for a single permutation of the data
}
\description{
Single permutation

Single permutation statistic of \code{ulsif} object

Single permutation statistic of \code{kliep} object

Single permutation statistic of \code{lhss} object

Single permutation statistic of \code{spectral} object

Single permutation statistic of \code{naivedensityratio} object
}
