% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_univariate}
\alias{plot_univariate}
\title{Scatter plot of density ratios and individual variables}
\usage{
plot_univariate(
  object,
  vars,
  samples = "both",
  logscale = TRUE,
  output = "individual",
  facet = FALSE,
  nrow = NULL
)
}
\arguments{
\item{object}{Density ratio object created with e.g., \code{\link[=kliep]{kliep()}}, \code{\link[=ulsif]{ulsif()}},
or \code{\link[=naive]{naive()}}}

\item{vars}{Character vector of variable names to be plotted.}

\item{logscale}{Logical indicating whether to plot the density ratio
estimates on a log scale. Default is TRUE.}
}
\description{
Plot a scatter plot showing the relationship between estimated densityratios and individual variables.Said differently, displays which densityratios are more likely for which values of the individual variables.
}
