% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_univariate}
\alias{plot_univariate}
\title{Scatter plot of density ratios and individual variables}
\usage{
plot_univariate(
  object,
  vars = NULL,
  samples = "both",
  logscale = TRUE,
  grid = FALSE,
  sample.facet = FALSE,
  nrow = NULL,
  tol = 0.01,
  ...
)
}
\arguments{
\item{object}{Density ratio object created with e.g., \code{\link[=kliep]{kliep()}}, \code{\link[=ulsif]{ulsif()}},
or \code{\link[=naive]{naive()}}}

\item{vars}{Character vector of variable names to be plotted.}

\item{samples}{Character string indicating whether to plot the 'numerator',
'denominator', or 'both' samples. Default is 'both'.}

\item{logscale}{}

\item{grid}{Logical indicating whether output should be a list of individual plots ("individual"), or one facetted plot with all variables ("assembled"). Defaults to "individual".}

\item{sample.facet}{Logical indicating whether to facet the plot by sample, i.e, showing plots separate for each sample, and side to side. Defaults to FALSE.}

\item{nrow}{Integer indicating the number of rows in the assembled plot. If NULL, the number of rows is automatically calculated.}

\item{tol}{Numeric indicating the tolerance: values below this value will be set to the tolerance value, for legibility of the plots}

\item{...}{Additional arguments passed to the predict() function.}
}
\value{
Scatter plot of density ratios and individual variables.
}
\description{
A scatter plot showing the relationship between estimated density ratios and individual variables.
}
