% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.density}
\alias{predict.density}
\title{Predict function for density object}
\usage{
\method{predict}{density}(object, newdata, lambda = 1e-09, log = FALSE)
}
\description{
Predict function for density object
}
\keyword{internal}
