% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.lhss}
\alias{predict.lhss}
\title{Obtain predicted density ratio values from a \code{lhss} object}
\usage{
\method{predict}{lhss}(
  object,
  newdata = NULL,
  sigma = c("sigmaopt", "all"),
  lambda = c("lambdaopt", "all"),
  ...
)
}
\arguments{
\item{object}{A \code{lhss} object}

\item{newdata}{Optional \code{matrix} new data set to compute the density}

\item{sigma}{A scalar with the Gaussian kernel width}

\item{lambda}{A scalar with the regularization parameter}

\item{...}{Additional arguments to be passed to the function}
}
\value{
An array with predicted density ratio values from possibly new data,
but otherwise the numerator samples.
}
\description{
Obtain predicted density ratio values from a \code{lhss} object
}
\examples{
x <- rnorm(100) |> matrix(50)
y <- rnorm(200, 1, 2) |> matrix(100)
fit1 <- lhss(x, y, m = 1)
predict(fit1)
predict(fit1, newdata = rbind(x, y))
predict(fit1, newdata = rbind(x, y), sigma = 2)
}
\seealso{
\code{\link{predict}}, \code{\link{lhss}}
}
\keyword{lhss}
\keyword{predict}
