% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ulsif}
\alias{print.ulsif}
\alias{print.summary.ulsif}
\alias{print.kliep}
\alias{print.naivedensityratio}
\alias{print.naivesubspacedensityratio}
\title{Print a \code{ulsif} object}
\usage{
\method{print}{ulsif}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{summary.ulsif}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{kliep}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{naivedensityratio}(x, digits = max(3L, getOption("digits") - 3L), ...)

\method{print}{naivesubspacedensityratio}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{ulsif}, \code{summary.ulsif}, \code{kliep}
or \code{summary.kliep}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{ulsif} object.

\code{NULL}

\code{NULL}

\code{invisble} The inputted \code{naivedensityratio} object.

\code{invisble} The inputted \code{naivesubspacedensityratio} object.
}
\description{
Print a \code{ulsif} object

Print a \code{summary.ulsif} object

Print a \code{kliep} object

Print a \code{naivedensityratio} object

Print a \code{naivesubspacedensityratio} object
}
