% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.ulsif}
\alias{print.summary.ulsif}
\title{Print a \code{summary.ulsif} object}
\usage{
\method{print}{summary.ulsif}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{Object of class \code{summary.ulsif}.}

\item{digits}{Number of digits to use when printing the output.}

\item{...}{further arguments on how to format the number of digits.}
}
\value{
\code{invisble} The inputted \code{summary.ulsif} object.
}
\description{
Print a \code{summary.ulsif} object
}
\seealso{
\code{\link{print}}, \code{\link{summary.ulsif}}, \code{\link{ulsif}}
}
