% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment.R
\name{recruitment}
\alias{recruitment}
\title{Recruitment}
\usage{
recruitment(t, cr, nm, SSB, A = 5, R0 = 1e+05, H, B0, Eps, Sigma_R,
  Rec_age, Recruitment_mode, LDP = 0.1)
}
\arguments{
\item{t}{temporary numeric value, the current time step .}

\item{cr}{temporary numeric value, the current control rule .}

\item{nm}{temporary numeric value, the current natural mortality estimate.}

\item{SSB}{numeric array, the spawning stock biomass of the whole stock for
each area, at each timestep, under each control rule, and for each
estimate of natural mortality, in kg.}

\item{A}{numeric value, the number of total areas in the model. Default
value is 5.}

\item{R0}{numeric value, set arbitrarily, the unfished recruitment. Default
value is 1e+5.}

\item{H}{numeric value, the steepness of the stock-recruitment curve.}

\item{B0}{numeric value, the unfished biomass, in kg.}

\item{Eps}{numeric matrix, the recruitment error terms.}

\item{Sigma_R}{numeric value, the recruitment standard deviation.}

\item{Rec_age}{numeric value, the age at recruitment, in years.}

\item{Recruitment_mode}{character value, values can be:
'closed' - the recruits in each area originate from adults in that area.
'pool' - the recruits in each area come from a pool of larvae produced by
   adults in all areas.
Default value is 'pool'.}

\item{LDP}{numeric value, the larval drift proportion, the proportion of
larvae that drift from one area to an adjacent area before settling.
Default value is 0.1.}
}
\value{
a numeric value representing the number of new recruits coming into
   the population in area a, at timestep t, under control rule cr, with an
   estimate of natural mortality of nm.
}
\description{
\code{recruitment} calculates the number of new recruits entering the
   population
}
\examples{
SSB <- array(rep(10, 5*70*6*3), c(5, 70, 6, 3))
NuR <- array(rnorm(5*70*6*3, 0, 0.5), c(5, 70, 6, 3))
Eps <- epsilon(A = 5, TimeT = 70, CR = 6, NM = 3, NuR, Rho_R = 0)
recruitment(t = 3, cr = 1, nm = 2, SSB, A = 5, R0 = 1e+5, H = 0.65,
   B0 = 1e+5/1.1, Eps, Sigma_R = 0.5, Rec_age = 2, Recruitment_mode = 'pool',
   LDP = 0.1)
}
