% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ulsif.R
\name{ulsif}
\alias{ulsif}
\title{Unconstrained least-squares importance fitting}
\usage{
ulsif(nu, de, sigma = NULL, lambda = NULL, ncenters = nrow(nu), centers = NULL)
}
\arguments{
\item{nu}{Numeric matrix with numerator samples}

\item{de}{Numeric matrix with denominator samples (must have the same
variables as \code{nu})}

\item{sigma}{\code{NULL} or a scalar value to determine the bandwidth of the
Gaussian kernel gram matrix. If \code{NULL}, sigma is the median Euclidean
interpoint distance.}

\item{lambda}{\code{NULL} or a scalar value to determine the regularization
imposed on the Gaussian kernel gram matrix of the denominator samples. If
\code{NULL}, \code{lambda} is chosen to be \eqn{\sqrt{N}}.}

\item{ncenters}{Maximum number of Gaussian centers in the kernel gram
matrix. Defaults to all numerator samples.}

\item{centers}{Numeric matrix with the same variables as \code{nu} and
\code{de} that are used as Gaussian centers in the kernel Gram matrix. By
default, the matrix \code{nu} is used as the matrix with Gaussian centers.}
}
\value{
\code{ulsif} returns \code{rhat}, the estimated density ratio.
}
\description{
Unconstrained least-squares importance fitting
}
\examples{
set.seed(1)
x <- rnorm(100) |> matrix(100)
y <- rnorm(200, 1, 2) |> matrix(200)
ulsif(x, y)
ulsif(x, y, sigma = 2, lambda = 2)
}
